requirements.txt
# Core
pandas>=2.0
numpy>=1.24
pyyaml>=6.0
python-dotenv>=1.0

# LLM clients
ollama>=0.4
openai>=1.0

# Metrics
scikit-learn>=1.3

# API
fastapi>=0.110
uvicorn>=0.30
pydantic>=2.0

# Experiment pipeline
tqdm>=4.66

# Analysis & visualization
matplotlib>=3.8
seaborn>=0.13
scipy>=1.11

# Frontend
streamlit>=1.38
